Update README.md
My idea will change the blockchain space and will bring blockchain to the real world.
I have an idea thats worth listening to because I'm not looking to own anything I want everything to be open source and the money thats made I want it to go to the people developing this. 
Basically I want to use the blockchain to incentivies people to share other peoples content through subscription from link hosting. The idea is to have people get paid monthly from sharing a link(link hosting) with an autogenerated link to a subscription service. When someone click the link host's link that person will have the option to subscribe to a content creators channel or service. The link host will take a cut of what consumer pays from the content creator and anyone can become a link host.  If the consumer subscribes to a content creators channel or service that link host gets a small monthly cut for each consumer they bring in. So if your cut was $1 per consumer as a link host, 10,000 people clicked your autogenerated link with them subscribing, then you get paid $10,000monthly. Also if someone unsubscribes the link host then blockchain records that, the link host will get $1 less from each person that unsubscribes. Essentially what you're doing is creating a smart contract for everyone to work together. Theres def more to this but my brain is fried right now but if you see potential in this just reach out to me so I can really share with you over the phone my vision and maybe you can bring this vision to life.